// bemfeito_public_new/ClientApp/app/models/trader.profile.interface.ts
export interface ITraderProfile {
    label: string;
    name: string;
    telephone: string;
    email: string;
    website: any;
    profiles: any;
    addresses: any;
    contacts: any;
    categories: any;
    notes: any;
    gallery: any;
    reviews: any;
    reviewStatistic: any;
    traderQualifications: any;
    insuranceDocuments: any;
    userId: any;
    user: any;
    id: any;

}